add db-users tests
