+++ src/spec/pocky-db-spec.ts
@@ -2,8 +2,6 @@ import PockyDB from '../lib/database/pocky-db';
 import Config from '../lib/config';
 import { QueryConfig, Client, QueryResult } from 'pg';
 import { Role } from '../models/database';
-import { CiscoSpark } from 'ciscospark/env';
-import MockCiscoSpark from './mocks/mock-spark';
 import MockConfig from './mocks/mock-config';
 import QueryHandler from '../lib/database/query-handler-interface';
 import MockQueryHandler from './mocks/mock-query-handler';
@@ -37,79 +35,6 @@ beforeAll(() => {
 		throw new Error('bad config');
 	});
 });
-
-function createPgClient(connectSuccess : boolean, pegCount : number) : Client {
-	let client = new Client();
-
-	if (connectSuccess) {
-		console.log('connect success');
-		spyOn(client, 'connect').and.returnValue(new Promise((resolve, reject) => resolve()));
-	} else {
-		console.log('connect fail');
-		spyOn(client, 'connect').and.returnValue(new Promise((resolve, reject) => reject()));
-	}
-
-	spyOn(client, 'query').and.callFake((statement : QueryConfig) => {
-		switch(statement.name) {
-			case 'returnResultsQuery':
-				return new Promise((resolve, reject) => {
-					resolve({rows: 'mock name'});
-				});
-			case 'returnWinnersQuery':
-				expect(statement.values[0]).toBe(5);
-				expect(statement.values[1]).toBe(3);
-				return new Promise((resolve, reject) => {
-					resolve({rows: 'mock name'});
-				});
-			case 'resetQuery':
-				return new Promise((resolve, reject) => {
-					resolve('reset return');
-				});
-			case 'createUserQuery':
-				expect(statement.values[0]).toBe('some_sender');
-				return new Promise((resolve, reject) => {
-					resolve('create return');
-				});
-			case 'givePegWithCommentQuery':
-				expect(statement.values[0]).toBe('some_sender');
-				expect(statement.values[1]).toBe('some_receiver');
-				expect(statement.values[2]).toBe('some comment here');
-				return new Promise((resolve, reject) => {
-					resolve();
-				});
-			case 'existsQuery':
-				expect(statement.values[0] === 'some_sender' || statement.values[0] === 'some_receiver').toBe(true);
-				return new Promise((resolve, reject) => {
-					resolve({
-						rows: [{exists:true}]
-					});
-				});
-			case 'pegsGiven':
-				expect(statement.values[0]).toBe('some_sender');
-				return new Promise((resolve, reject) => {
-					resolve({
-						rows: [{count:pegCount}]
-					});
-				});
-		}
-	});
-
-	return client;
-}
-
-function createSparkMock() : CiscoSpark {
-	let spark = new MockCiscoSpark();
-
-	spyOn(spark.people, 'get').and.callFake((userid : string) => {
-		return new Promise((resolve, reject) => {
-			resolve({
-				displayName: userid + 'display'
-			});
-		});
-	});
-
-	return spark;
-}
 
 function createQueryHandlerMock(result : any | QueryResult) : QueryHandler {
 	let queryHandler = new MockQueryHandler(result);
